+++ config/backpack/ui.php
@@ -95,7 +95,6 @@ return [
     // CSS files that are loaded in all pages, using Laravel's asset() helper
     'styles' => [
         // 'styles/example.css',
-        // 'https://some-cdn.com/example.css',
     ],
 
     // CSS files that are loaded in all pages, using Laravel's mix() helper
@@ -116,9 +115,6 @@ return [
     // JS files that are loaded in all pages, using Laravel's asset() helper
     'scripts' => [
         // 'js/example.js',
-        // 'https://cdn.jsdelivr.net/npm/vue@2.4.4/dist/vue.min.js',
-        // 'https://cdn.jsdelivr.net/npm/react@16/umd/react.production.min.js',
-        // 'https://cdn.jsdelivr.net/npm/react-dom@16/umd/react-dom.production.min.js',
     ],
 
     // JS files that are loaded in all pages, using Laravel's mix() helper